Update input file naming
When you save the AoC input for a given day (at least from your browser), it will default to saving as a .txt file. Currently, my file reading assumes that there is no file extension, so you have to rename the file from `input.txt` to `day_X`. This can be a tiny bit annoying, as at least on MacOS it defaults to selecting just the file name and not the extension when you rename.

To address this, update the file reading logic to look for `day_X.txt` instead.